refactor: delete SASS dep, use native CSS
fixes https://github.com/TACC/APCD-CMS/security/dependabot/42
