+++ apcd_cms/src/client/package.json
@@ -75,7 +75,6 @@
     "react-test-renderer": "^17.0.2",
     "redux-mock-store": "^1.5.3",
     "redux-saga-test-plan": "^4.0.0-rc.3",
-    "sass": "^1.42.1",
     "stylelint": "^14.4.0",
     "stylelint-config-recommended": "^7.0.0",
     "stylelint-config-standard": "^25.0.0",
