+++ apcd_cms/src/client/src/components/Admin/ViewUsers/ViewUsers.module.css
@@ -1,8 +1,8 @@
-$modal-width: 55%;
-
 .customModal {
-  max-width: $modal-width !important;
-  width: $modal-width;
+  --width: 55%;
+
+  max-width: var(--width) !important;
+  width: var(--width);
 }
 
 .modal-header {
@@ -12,18 +12,18 @@ $modal-width: 55%;
 }
 
 .customModalTitle {
-  text-transform: none !important; // Prevent text from being all caps
+  text-transform: none !important; /* Prevent text from being all caps */
   font-size: 2.5rem;
-  margin: 0; // Ensure there's no margin
+  margin: 0; /* Ensure there's no margin */
 }
 
 .customLabel {
-  font-size: 1.5rem; // Adjust the font size as needed
+  font-size: 1.5rem; /* Adjust the font size as needed */
 }
 
 .customSubmitButton {
-  font-size: 1.5rem; // Increase font size
-  padding: 0.75rem 1.5rem; // Increase padding
+  font-size: 1.5rem; /* Increase font size */
+  padding: 0.75rem 1.5rem; /* Increase padding */
 }
 
 .viewRecord {
